third_party/blink/web_tests/platform/mac/external/wpt/html/browsers/windows/clear-window-name.https-expected.txt

This is a testharness.js-based test.
PASS Window.name is not reset when there is an opener around
PASS Window.name is not reset when there is an opener around (cross-origin)
PASS Window.name is not reset at the first navigation away from initial about:blank with noopener
FAIL Window.name is reset at the first cross-origin navigation with noopener assert_equals: expected "" but got "5c9611a5-e210-4535-a99d-5edd35cd8297"
PASS Window.name is not reset at the first navigation away from initial about:blank with window.opener set to null
PASS Window.name is not reset at the same-origin navigation with window.opener set to null
FAIL Window.name is reset at the first first cross-origin navigation with window.opener set to null assert_equals: expected "" but got "251019f0-8b9b-4998-8c44-f237354d528b"
PASS Window.name is set by the window
FAIL Window.name is reset at the first cross-origin navigation assert_equals: expected "" but got "1e59a1a2-bfe5-4bfb-a91d-21c550e72ae3"
Harness: the test ran to completion.
